+++ src/geometry-utils.ts
@@ -118,7 +118,8 @@ export class GeometryUtils {
     let i = 0;
     for (const geometry of geometries) {
       const size = geometry.attributes.position.count;
-      const array = new Uint8Array(size).fill(i++);
+      // TODO: Substitute blockID attribute by block id map
+      const array = new Uint16Array(size).fill(i++);
       geometry.setAttribute("blockID", new BufferAttribute(array, 1));
     }
   }
